requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
fpdf==1.7.2
python-dotenv==1.0.0
